remove unecessary library, refactor image picker service
